Move carefluence and techcare webscrapers out of endpoint list querier and into chpl endpoint list querier
