+++ endpointmanager/pkg/endpointwebscraper/endpointwebscraper.go
@@ -17,17 +17,11 @@ type LanternEntry struct {
 
 var oneUpURL = "https://1up.health/fhir-endpoint-directory"
 var careEvolutionURL = "https://fhir.docs.careevolution.com/overview/public_endpoints.html"
-var techCareURL = "https://devportal.techcareehr.com/Serviceurls"
-var carefluenceURL = "https://carefluence.com/carefluence-fhir-endpoints/"
 
 func EndpointListWebscraper(vendorURL string, vendor string, fileToWriteTo string) {
 
 	if vendorURL == oneUpURL || vendorURL == careEvolutionURL {
 		HTMLtablewebscraper(vendorURL, vendor, fileToWriteTo)
-	} else if vendorURL == techCareURL {
-		Techcarewebscraper(vendorURL, fileToWriteTo)
-	} else if vendorURL == carefluenceURL {
-		Carefluenceebscraper(vendorURL, fileToWriteTo)
 	}
 }
